refact: project_empresa, project_bairro and history_project refatored

+++ src/infra/relational/repository/project_empresa_repository.py
@@ -1,18 +1,20 @@
 from typing import List
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy import and_
 
 from src.domain.entities.project_empresa import ProjectEmpresaEntity
 from src.infra.relational.models.project_empresa import ProjectEmpresa
 from src.infra.relational.config.interface.i_db_connection_handler import IDBConnectionHandler
-
-from src.errors.repository.already_exists_error.project_empresa_already_exists import ProjectEmpresaAlreadyExists
-from src.errors.repository.not_exists_error.projects_from_empresa_does_not_exists import ProjectsFromEmpresaDoesNotExists
-from src.errors.repository.error_on_delete.error_on_update_empresa_from_project import ErrorOnUpdateEmpresaFromProject
-from src.errors.repository.error_on_delete.error_on_delete_project_empresa import ErrorOnDeleteProjectEmpresa
-
 from src.data.interface.i_project_empresa_repository import IProjectEmpresaRepository
 
+# Errors
+from src.errors.repository.already_exists_error.project_empresa_already_exists import ProjectEmpresaAlreadyExists
+from src.errors.repository.not_exists_error.project_empresa_not_exists import ProjectEmpresaNotExists
+from src.errors.repository.error_on_delete.error_on_update_empresa_from_project import ErrorOnUpdateEmpresaFromProject
+from src.errors.repository.error_on_delete.error_on_delete_project_empresa import ErrorOnDeleteProjectEmpresa
+from src.errors.repository.error_on_insert.error_on_insert_project_empresa import ErrorOnInsertProjectEmpresa
+from src.errors.repository.error_on_find.error_on_find_project_empresa import ErrorOnFindProjectEmpresa
+from src.errors.repository.has_related_children.project_empresa_has_related_children import ProjectEmpresaHasRelatedChildren
+from sqlalchemy.exc import IntegrityError
 
 class ProjectEmpresaRepository(IProjectEmpresaRepository):
 
@@ -29,7 +31,7 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                 message=f'Association (project_id={project_id}, empresa_id={empresa_id}) already exists.'
             ) from e
         except Exception as e:
-            raise e
+            raise ErrorOnInsertProjectEmpresa(message=f'Error on insert project empresa project_id={project_id}, empresa_id={empresa_id}: {str(e)}') from e
 
     def find(self, project_id: int, empresa_id: int) -> ProjectEmpresaEntity:
         try:
@@ -41,7 +43,7 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     )
                 ).first()
                 if not assoc:
-                    raise ProjectsFromEmpresaDoesNotExists(
+                    raise ProjectEmpresaNotExists(
                         message=f'No association found for project_id={project_id} and empresa_id={empresa_id}'
                     )
                 return ProjectEmpresaEntity(
@@ -49,8 +51,10 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     empresa_id=assoc.empresa_id,
                     created_at=assoc.created_at
                 )
+        except ProjectEmpresaNotExists as e:
+            raise e from e
         except Exception as e:
-            raise e
+            raise ErrorOnFindProjectEmpresa(message=f'Error on find project empresa project_id={project_id}, empresa_id={empresa_id}') from e
 
     def find_all_from_empresa(self, empresa_id: int) -> List[ProjectEmpresaEntity]:
         try:
@@ -59,7 +63,7 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     ProjectEmpresa.empresa_id == empresa_id
                 ).all()
                 if not associations:
-                    raise ProjectsFromEmpresaDoesNotExists(
+                    raise ProjectEmpresaNotExists(
                         message=f'No projects found for empresa_id={empresa_id}'
                     )
                 return [
@@ -70,8 +74,10 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     )
                     for assoc in associations
                 ]
+        except ProjectEmpresaNotExists as e:
+            raise e from e
         except Exception as e:
-            raise e
+            raise ErrorOnFindProjectEmpresa(message=f'Error on find project empresa empresa_id={empresa_id}') from e
 
     def update_empresa(self, project_id: int, empresa_id: int, new_empresa_id: int) -> None:
         try:
@@ -83,6 +89,8 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     )
                 ).update({'empresa_id': new_empresa_id})
                 db.session.commit()
+        except IntegrityError as e:
+            raise ProjectEmpresaAlreadyExists(f'Project empresa with project_id={project_id}, empresa_id={new_empresa_id} already exists: {str(e)}') from e
         except Exception as e:
             raise ErrorOnUpdateEmpresaFromProject(
                 message=f'Error updating project {project_id} from empresa {empresa_id} to {new_empresa_id}'
@@ -91,6 +99,14 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
     def delete(self, project_id: int, empresa_id: int) -> None:
         try:
             with self.__db_connection_handler as db:
+                project_empresa = db.session.query(ProjectEmpresa).filter(
+                    and_(
+                        ProjectEmpresa.project_id == project_id,
+                        ProjectEmpresa.empresa_id == empresa_id
+                    )
+                ).first()
+                if not project_empresa:
+                    raise ProjectEmpresaNotExists(message=f'Project empresa project_id={project_id}, empresa_id={empresa_id} does not exists')
                 db.session.query(ProjectEmpresa).filter(
                     and_(
                         ProjectEmpresa.project_id == project_id,
@@ -98,6 +114,10 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
                     )
                 ).delete()
                 db.session.commit()
+        except IntegrityError as e:
+            raise ProjectEmpresaHasRelatedChildren(message=f'Project empresa project_id={project_id}, empresa_id={empresa_id} has related children: {str(e)}') from e
+        except ProjectEmpresaAlreadyExists as e:
+            raise e from e
         except Exception as e:
             raise ErrorOnDeleteProjectEmpresa(
                 message=f'Error deleting association (project_id={project_id}, empresa_id={empresa_id}): {e}'
@@ -106,10 +126,21 @@ class ProjectEmpresaRepository(IProjectEmpresaRepository):
     def delete_all_from_project(self, project_id: int) -> None:
         try:
             with self.__db_connection_handler as db:
+                project_empresa = db.session.query(ProjectEmpresa).filter(
+                    and_(
+                        ProjectEmpresa.project_id == project_id
+                    )
+                ).all()
+                if not any(project_empresa):
+                    raise ProjectEmpresaNotExists(message=f'Project empresa from project project_id={project_id} does not exists')
                 db.session.query(ProjectEmpresa).filter(
                     ProjectEmpresa.project_id == project_id,
                 ).delete()
                 db.session.commit()
+        except IntegrityError as e:
+            raise ProjectEmpresaHasRelatedChildren(message=f'Project empresa from project project_id={project_id} has related children: {str(e)}') from e
+        except ProjectEmpresaNotExists as e:
+            raise e from e
         except Exception as e:
             raise ErrorOnDeleteProjectEmpresa(
                 message=f'Error deleting association (project_id={project_id}): {e}'
